12/29/2017

+++ app/Classes/ImageUploader.php
@@ -14,13 +14,14 @@
   class ImageUploader
   {
 
-    function __construct($input, $obj, $disk, $path)
+    function __construct($input, $obj, $dbcolumn, $disk, $path)
     {
       $this->input = $input;
       $this->obj = $obj;
+      $this->dbcolumn = $dbcolumn;
       $this->storage = Storage::disk($disk);
       $this->img = Image::make($this->input);
-      $this->imgName = date('ymd'.time()).str_shuffle ('accounts').".jpg";
+      $this->imgName = date('ymd'.time()).$this->generateRandomString().".jpg";
       $this->imgPath = $path;
       $this->imgDbPath = $this->storage->url($path).$this->imgName;
     }
@@ -28,11 +29,17 @@
     public function removeOld()
     {
       // Find Stored URL
-      $storedImage = $this->obj->avatar;
-      // Explode to get Path
-      $storedImage = explode('.com/', $storedImage);
+      $storedImage = $this->obj[$this->dbcolumn];
+
       // Delete from Storage
-      $this->storage->delete($storedImage[1]);
+      $db = explode('/', 'https://coviweb.s3.amazonaws.com/images/accounts_logos/1712291514582636Q1nczks2RD.jpg');
+      $s3 = explode('/', Storage::url('/'));
+
+      if ($db[2] == $s3[2]) {
+        // Explode to get Path
+        $storedImage = explode('.com/', $storedImage);
+        $this->storage->delete($storedImage[1]);
+      }
     }
 
     public function store($size = 100)
@@ -43,7 +50,7 @@
       $this->storage->put($this->imgPath.$this->imgName, $this->img->stream('jpg', 60)->__toString());
       // Update DB Path
       $this->obj->update([
-        'avatar' => $this->imgDbPath
+        $this->dbcolumn => $this->imgDbPath
       ]);
     }
 
@@ -58,6 +65,10 @@
       $width = $this->img->width();
       // Create a canvas with white background to fix PNG black bg
       $this->img = Image::canvas($width, $heigth, '#fefefe')->insert($this->img, 'center');
+    }
+
+    public function generateRandomString($length = 10) {
+        return substr(str_shuffle(str_repeat($x='0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', ceil($length/strlen($x)) )),1,$length);
     }
 
   }
